requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy<2.0.0  # Force numpy 1.x for compatibility
scikit-learn>=0.24.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.3.0
opencv-python>=4.5.0
tqdm>=4.62.0
pyyaml>=5.4.0
tensorboard>=2.7.0
timm>=0.6.0  # Add timm for EfficientNet